requirements.txt
cassandra-driver
kafka-python
praw
apache-airflow
pyspark
confluent-kafka 